Its finally useful.

+++ app/src/main/java/com/app/vwflashtools/LoggingFragment.kt
@@ -6,7 +6,6 @@ import android.content.Intent
 import android.content.IntentFilter
 import android.graphics.Color
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -227,7 +226,7 @@ class LoggingFragment : Fragment() {
                                     //Write new values to log
                                     val bleHeader = BLEHeader()
                                     bleHeader.fromByteArray(readBuff)
-                                    var strItems: String? = (bleHeader.tickCount / 1000).toString()
+                                    var strItems: String? = (bleHeader.tickCount.toFloat() / 1000.0f).toString()
                                     for (i in 0 until dList.count()) {
                                         strItems += ",${DIDList[dList[i].toInt()].value}"
                                     }
